# assignment-4/orders/openapi.yaml
openapi: 3.0.3

info:
  title: CampusEats Orders Service
  description: REST API for managing CampusEats orders.
  version: 1.0.0

servers:
  - url: http://localhost:5000

paths:

  /orders:
    post:
      summary: Create a new order
      operationId: createOrder

      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
          description: Unique key used to safely retry order creation.

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'

      responses:
        '201':
          description: Order created successfully.
          headers:
            Location:
              description: URL of the newly created order.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

        '400':
          description: Malformed request body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'

        '409':
          description: Idempotency key conflict or duplicate request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'

        '422':
          description: Request is valid but the domain refuses the order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'

        '503':
          description: Payment service unavailable after retries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'


    get:
      summary: List orders
      operationId: listOrders

      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/OrderStatus'
          description: Filter orders by their current status.

      responses:
        '200':
          description: List of matching orders.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponse'

        '400':
          description: Invalid status filter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'


  /orders/{id}:
    get:
      summary: Get an order
      operationId: getOrder

      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique order identifier.

      responses:
        '200':
          description: Order found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

        '404':
          description: Order not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'


  /orders/{id}/cancellation:
    post:
      summary: Cancel an order
      operationId: cancelOrder

      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique order identifier.

      responses:
        '200':
          description: Order cancelled successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'

        '404':
          description: Order not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'

        '409':
          description: Order cannot be cancelled in its current state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'


components:

  schemas:

    CreateOrderRequest:
      type: object
      required:
        - studentId
        - addressId
        - items
      properties:
        studentId:
          type: string
          example: STU-1001

        addressId:
          type: string
          example: ADDR-101

        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'


    OrderItemRequest:
      type: object
      required:
        - itemId
        - quantity
      properties:
        itemId:
          type: string
          example: ITEM-101

        quantity:
          type: integer
          minimum: 1
          example: 2


    OrderResponse:
      type: object
      required:
        - orderId
        - studentId
        - addressId
        - items
        - total
        - status
      properties:
        orderId:
          type: string
          example: ORD-1001

        studentId:
          type: string
          example: STU-1001

        addressId:
          type: string
          example: ADDR-101

        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'

        total:
          type: number
          format: double
          example: 499.00

        status:
          $ref: '#/components/schemas/OrderStatus'


    OrderItemResponse:
      type: object
      required:
        - itemId
        - quantity
        - price
      properties:
        itemId:
          type: string
          example: ITEM-101

        quantity:
          type: integer
          example: 2

        price:
          type: number
          format: double
          example: 249.50


    OrderStatus:
      type: string
      enum:
        - PLACED
        - CONFIRMED
        - PREPARING
        - READY
        - OUT_FOR_DELIVERY
        - DELIVERED
        - CANCELLED


    Problem:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          example: https://api.campuseats.example.com/problems/order-not-found

        title:
          type: string
          example: Order not found

        status:
          type: integer
          example: 404

        detail:
          type: string
          example: No order exists with id ORD-9999.